fix: rapid fire off-by-one, destroy handler crashes, and 6 type-safety fixes
BattleEngine:
- GeometricDistribution::getProbabilityFromMean() used 1/\ instead of 1/(\+1)
  RF=200 was producing ~1 extra shot instead of 200. All ships with RF affected.

Destroy handler:
- \[Ships::ship_deathstar] undefined index crash (null coalesce)
- Division by zero when \ == 0 after battle (guard added)
- getShipType() missing (int) cast (matched Attack.php)

Type-safety (PHP 8.5 strict):
- Formulas: getMoonDestructionChance, getTearDownBaseCost, getTearDownCost round()/floor() float→int
- ShipyardController: getMissilesItemLimit floor() float→int (IPM build crash)
- Expedition: updateFleetEndTime float arg in resultDelay + resultEarly
- Attack: getAllAcsFleetsByGroupId string→int cast
- GalaxyController: \['galaxy/system/planet'] null coalesce
- Lang: added mission 10 (Missile) to EN + ES type_mission arrays

+++ legacy/app/Http/Controllers/Game/ShipyardController.php
@@ -471,7 +471,7 @@ class ShipyardController extends BaseController
         }
 
         if ($itemId == Defenses::defense_interplanetary_missile) {
-            $amount = floor($max_amount / 2);
+            $amount = (int) floor($max_amount / 2);
         }
 
         if ($amount_requested > $amount) {
